modify models and move to shared folder
