+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cpsc436-group-project",
-  "version": "0.19.0",
+  "version": "0.13.1",
   "private": true,
   "scripts": {
     "test": "nyc --reporter html --reporter json --reporter text npm run tests && nyc report --reporter=text-lcov > coverage.lcov",
