+++ src/backend/models/PostModel.ts
@@ -1,23 +1,12 @@
 import {Document, Model, model, Schema, Types} from "mongoose";
 import ObjectId = Types.ObjectId;
-import {ITag} from "./TagModel";
+import { IPost } from "../../shared/ModelInterfaces";
 
-interface IPostSchema extends Document {
-    _id: string,
-    time: Date,
-    userId: string,
-    detail: string,
-    uploadedFiles: IUploadedFile[],
-    type: string,
-    visibility: string,
-    tags: ITag[],
-    likedUserIds: string[]
+interface IPostSchema extends Document, IPost {
+    _id: string
 }
 
-export interface IPost extends IPostSchema {
-}
-
-interface IPostModel extends Model<IPost> {
+interface IPostModel extends Model<IPostSchema> {
 }
 
 const postSchema: Schema = new Schema({
@@ -34,9 +23,6 @@ const postSchema: Schema = new Schema({
     likedUserIds: [ObjectId]
 });
 
-export const Post = model<IPost, IPostModel>('post', postSchema);
+export const Post = model<IPostSchema, IPostModel>('post', postSchema);
 
-export interface IUploadedFile {
-    type: string,
-    path: string
-}
+
