+++ test/backend/api/v1/ChatTest.spec.ts
@@ -1,7 +1,8 @@
 import chai, {expect} from 'chai';
 const mongoose = require("mongoose");
 import chaiHttp = require('chai-http');
-import {Chat, IChat} from "../../../../src/backend/models/ChatModel";
+import {Chat} from "../../../../src/backend/models/ChatModel";
+import { IChat } from '../../../../src/shared/ModelInterfaces';
 import {User} from "../../../../src/backend/models/UserModel";
 
 chai.use(chaiHttp);
